Do refactor and add network service
