+++ app/build.gradle
@@ -45,8 +45,7 @@ android {
     productFlavors {
         serv_ {
             buildConfigField STRING, ENDPOINT, server
-            versionName "$appVersion" + buildString
-            versionNameSuffix "-server"
+            versionName "$appVersion"
             dimension DEFAULT_DIMEN
         }
     }
@@ -59,13 +58,13 @@ dependencies {
     implementation 'com.android.support:design:27.1.1'
     implementation 'com.android.support.constraint:constraint-layout:1.1.3'
     // Retrofit.
-    implementation 'com.squareup.retrofit2:retrofit:2.4.0'
-    implementation 'com.google.code.gson:gson:2.8.2'
-    implementation 'com.squareup.retrofit2:converter-gson:2.1.0'
+    implementation 'com.squareup.retrofit2:retrofit:2.5.0'
+    implementation 'com.google.code.gson:gson:2.8.5'
+    implementation 'com.squareup.retrofit2:converter-gson:2.5.0'
     // OKHttp.
-    implementation 'com.squareup.okhttp3:okhttp:3.10.0'
+    implementation 'com.squareup.okhttp3:okhttp:3.12.0'
 
-    implementation 'com.github.bumptech.glide:glide:4.8.0'
+    implementation 'com.github.bumptech.glide:glide:4.9.0'
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
 
@@ -75,8 +74,4 @@ dependencies {
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
     implementation 'com.android.support:cardview-v7:27.1.1'
-}
-
-static def getBuildString() {
-    "|${new Date().format('yy.MM.dd-HH.mm').toString()}"
 }
