--- app/src/androidTest/java/com/weatherkremenchug/weathermain/ExampleInstrumentedTest.kt
@@ -1,25 +1,0 @@
-package com.weatherkremenchug.weathermain
-
-import android.support.test.InstrumentationRegistry
-import android.support.test.runner.AndroidJUnit4
-
-import org.junit.Test
-import org.junit.runner.RunWith
-
-import org.junit.Assert.*
-
-/**
- * Instrumented test, which will execute on an Android device.
- *
- * @see [Testing documentation](http://d.android.com/tools/testing)
- */
-@RunWith(AndroidJUnit4::class)
-class ExampleInstrumentedTest {
-    @Test
-    fun useAppContext() {
-        // Context of the app under test.
-        val appContext = InstrumentationRegistry.getTargetContext()
-
-        assertEquals("com.weatherkremenchug.weatherkremenchugtest", appContext.packageName)
-    }
-}
