+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.weatherkremenchug.weathermain">
+    xmlns:tools="http://schemas.android.com/tools"
+    package="com.weather.weathermain">
 
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" />
@@ -12,9 +13,10 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/AppTheme">
+        android:theme="@style/AppTheme"
+        tools:ignore="GoogleAppIndexingWarning">
         <activity
-            android:name="com.weatherkremenchug.weathermain.activity.mainactivity.MainActivity"
+            android:name="com.weather.weathermain.activity.main.MainActivity"
             android:configChanges="orientation|screenSize"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar">
@@ -25,11 +27,11 @@
             </intent-filter>
         </activity>
         <activity
-            android:name="com.weatherkremenchug.weathermain.activity.weatherontoday.WeatherOnTodayActivity"
+            android:name="com.weather.weathermain.activity.weathertoday.WeatherOnTodayActivity"
             android:configChanges="orientation|screenSize"
             android:theme="@style/AppTheme.NoActionBar" />
         <activity
-            android:name="com.weatherkremenchug.weathermain.activity.settingsactivity.SettingsActivity"
+            android:name="com.weather.weathermain.activity.settings.SettingsActivity"
             android:label="@string/title_activity_settings"
             android:configChanges="orientation|screenSize"
             android:theme="@style/AppTheme.NoActionBar"/>
