+++ app/src/main/java/com/weather/weathermain/data/CloudsData.kt
@@ -1,0 +1,6 @@
+package com.weather.weathermain.data
+
+class CloudsData {
+
+    val all: String? = null
+}
