+++ app/src/main/java/com/weather/weathermain/data/CoordData.kt
@@ -1,0 +1,8 @@
+package com.weather.weathermain.data
+
+
+class CoordData {
+
+    val lat: String? = null
+    val lon: String? = null
+}
