+++ app/src/main/java/com/weather/weathermain/data/MainWeatherData.kt
@@ -1,0 +1,13 @@
+package com.weather.weathermain.data
+
+class MainWeatherData {
+
+    val temp: String? = null
+    val temp_min: String? = null
+    val temp_max: String? = null
+    val pressure: String? = null
+    val sea_level: String? = null
+    val grnd_level: String? = null
+    val humidity: String? = null
+    val temp_kf: String? = null
+}
