+++ app/src/main/java/com/weather/weathermain/data/RainData.kt
@@ -1,0 +1,3 @@
+package com.weather.weathermain.data
+
+class RainData
