+++ app/src/main/java/com/weather/weathermain/data/SysData.kt
@@ -1,0 +1,8 @@
+package com.weather.weathermain.data
+
+class SysData {
+//    val pod: String? = null
+//    val country: String? = null
+//    val sunrise: String? = null
+//    val sunset: String? = null
+}
