+++ app/src/main/java/com/weather/weathermain/data/WeatherData.kt
@@ -1,0 +1,9 @@
+package com.weather.weathermain.data
+
+class WeatherData {
+
+    val id: String? = null
+    val main: String? = null
+    val description: String? = null
+    val icon: String? = null
+}
