+++ app/src/main/java/com/weather/weathermain/data/WeatherForecast.kt
@@ -1,0 +1,10 @@
+package com.weather.weathermain.data
+
+import java.util.ArrayList
+
+class WeatherForecast {
+    val cod: String? = null
+    val message: String? = null
+    val cnt: String? = null
+    val list = ArrayList<ListData>()
+}
