+++ app/src/main/java/com/weather/weathermain/data/WindData.kt
@@ -1,0 +1,7 @@
+package com.weather.weathermain.data
+
+class WindData {
+
+    val speed: String? = null
+    val deg: String? = null
+}
