+++ app/src/main/java/com/weather/weathermain/data/enums/Enum.kt
@@ -1,0 +1,2 @@
+package com.weather.weathermain.data.enums
+
