+++ app/src/main/java/com/weather/weathermain/data/network/service/ServiceGenerator.kt
@@ -1,0 +1,12 @@
+package com.weather.weathermain.data.network.service
+
+class ServiceGenerator {
+    companion object {
+        var endPoint: String = ""
+    }
+
+    inline fun <reified S> createService(endPoint: String): S {
+        Companion.endPoint = endPoint
+    }
+
+}
