+++ app/src/main/java/com/weather/weathermain/domain/repository/IWeatherRepository.kt
@@ -1,0 +1,9 @@
+package com.weather.weathermain.domain.repository
+
+import com.weather.weathermain.data.repository.WeatherRepository
+
+interface IWeatherRepository {
+    companion object {
+        fun getInstance(): IWeatherRepository = WeatherRepository()
+    }
+}
