+++ app/src/main/java/com/weather/weathermain/utils/constants/Const.kt
@@ -1,0 +1,5 @@
+package com.weather.weathermain.utils.constants
+
+// MainActivity
+const val APP_ID = "c3eebf803f44713f50e31a7c5b215a73"
+const val APP_PREFERENCES = "settings"
