+++ app/src/main/java/com/weather/weathermain/utils/extensions/ViewExt.kt
@@ -1,0 +1,8 @@
+package com.weather.weathermain.utils.extensions
+
+import android.content.Context
+import android.widget.Toast
+
+fun Context.showToast(text: CharSequence) {
+    Toast.makeText(this, text, Toast.LENGTH_LONG).show()
+}
