--- app/src/main/java/com/weatherkremenchug/weathermain/GPSTracker.kt
@@ -1,185 +1,0 @@
-package com.weatherkremenchug.weathermain
-
-import android.Manifest
-import android.app.Activity
-import android.app.Service
-import android.content.Context
-import android.content.Intent
-import android.content.pm.PackageManager
-import android.location.Location
-import android.location.LocationListener
-import android.location.LocationManager
-import android.os.Bundle
-import android.os.IBinder
-import android.provider.Settings
-import android.support.v4.app.ActivityCompat
-import android.support.v4.content.ContextCompat
-import android.support.v7.app.AlertDialog
-import android.util.Log
-
-class GPSTracker(private val mContext: Activity) : Service(), LocationListener {
-    // flag for GPS status
-    internal var isGPSEnabled = false
-    // flag for network status
-    internal var isNetworkEnabled = false
-    // flag for GPS status
-    internal var canGetLocation = false
-
-    internal var location: Location? = null // location
-    internal var latitude: Double = 0.toDouble() // latitude
-    internal var longitude: Double = 0.toDouble() // longitude
-    // Declaring a Location Manager
-    protected var locationManager: LocationManager? = null
-
-    init {
-        if (ContextCompat.checkSelfPermission(mContext, android.Manifest.permission.ACCESS_FINE_LOCATION) == PackageManager.PERMISSION_GRANTED) {
-            getLocation()
-        } else {
-            ActivityCompat.requestPermissions(mContext, arrayOf(android.Manifest.permission.ACCESS_FINE_LOCATION), 1)
-        }
-
-    }
-
-    fun getLocation(): Location? {
-        try {
-            locationManager = mContext.getSystemService(Context.LOCATION_SERVICE) as LocationManager
-            // getting GPS status
-            //assert(locationManager != null)
-            isGPSEnabled = locationManager!!.isProviderEnabled(LocationManager.GPS_PROVIDER)
-            // getting network status
-            isNetworkEnabled = locationManager!!
-                    .isProviderEnabled(LocationManager.NETWORK_PROVIDER)
-            if (!isGPSEnabled && !isNetworkEnabled) {
-                // no network provider is enabled
-            } else {
-                this.canGetLocation = true
-                // First get location from Network Provider
-                if (isNetworkEnabled) {
-                    //check the network permission
-                    if (ActivityCompat.checkSelfPermission(mContext,
-                                    Manifest.permission.ACCESS_FINE_LOCATION) != PackageManager.PERMISSION_GRANTED && ActivityCompat.checkSelfPermission(mContext, Manifest.permission.ACCESS_COARSE_LOCATION) != PackageManager.PERMISSION_GRANTED) {
-                        ActivityCompat.requestPermissions(mContext,
-                                arrayOf(android.Manifest.permission.ACCESS_FINE_LOCATION, Manifest.permission.ACCESS_COARSE_LOCATION), 101)
-                    }
-                    locationManager!!.requestLocationUpdates(
-                            LocationManager.NETWORK_PROVIDER,
-                            MIN_TIME_BW_UPDATES,
-                            MIN_DISTANCE_CHANGE_FOR_UPDATES.toFloat(), this)
-                    Log.d("Network", "Network")
-                    if (locationManager != null) {
-                        location = locationManager!!
-                                .getLastKnownLocation(LocationManager.NETWORK_PROVIDER)
-                        if (location != null) {
-                            latitude = location!!.latitude
-                            longitude = location!!.longitude
-                        }
-                    }
-                }
-                // if GPS Enabled get lat/long using GPS Services
-                if (isGPSEnabled) {
-                    if (location == null) {
-                        //check the network permission
-                        if (ActivityCompat.checkSelfPermission(mContext, Manifest.permission.ACCESS_FINE_LOCATION) != PackageManager.PERMISSION_GRANTED && ActivityCompat.checkSelfPermission(mContext, Manifest.permission.ACCESS_COARSE_LOCATION) != PackageManager.PERMISSION_GRANTED) {
-                            ActivityCompat.requestPermissions(mContext,
-                                    arrayOf(android.Manifest.permission.ACCESS_FINE_LOCATION, Manifest.permission.ACCESS_COARSE_LOCATION), 101)
-                        }
-                        locationManager!!.requestLocationUpdates(
-                                LocationManager.GPS_PROVIDER,
-                                MIN_TIME_BW_UPDATES,
-                                MIN_DISTANCE_CHANGE_FOR_UPDATES.toFloat(), this)
-                        Log.d("GPS Enabled", "GPS Enabled")
-                        if (locationManager != null) {
-                            location = locationManager!!
-                                    .getLastKnownLocation(LocationManager.GPS_PROVIDER)
-                            if (location != null) {
-                                latitude = location!!.latitude
-                                longitude = location!!.longitude
-                            }
-                        }
-                    }
-                }
-            }
-        } catch (e: Exception) {
-            e.printStackTrace()
-        }
-
-        return location
-    }
-
-    /**
-     * Stop using GPS listener
-     * Calling this function will stop using GPS in your app
-     */
-    fun stopUsingGPS() {
-        if (locationManager != null) {
-            locationManager!!.removeUpdates(this@GPSTracker)
-        }
-    }
-
-    /**
-     * Function to get latitude
-     */
-    fun getLatitude(): Double {
-        getLocation()
-        if (location != null) {
-            latitude = location!!.latitude
-        }
-        // return latitude
-        return latitude
-    }
-
-    /**
-     * Function to get longitude
-     */
-    fun getLongitude(): Double {
-        if (location != null) {
-            longitude = location!!.longitude
-        }
-        // return longitude
-        return longitude
-    }
-
-    /**
-     * Function to check GPS/wifi enabled
-     * @return boolean
-     */
-    fun canGetLocation(): Boolean {
-        return this.canGetLocation
-    }
-
-    /**
-     * Function to show settings alert dialog
-     * On pressing Settings button will lauch Settings Options
-     */
-    fun showSettingsAlert() {
-        val alertDialog = AlertDialog.Builder(mContext)
-        // Setting Dialog Title
-        alertDialog.setTitle("GPS is settings")
-        // Setting Dialog Message
-        alertDialog.setMessage("GPS is not enabled. Do you want to go to settings menu?")
-        // On pressing Settings button
-        alertDialog.setPositiveButton("Settings") { dialog, which ->
-            val intent = Intent(Settings.ACTION_LOCATION_SOURCE_SETTINGS)
-            mContext.startActivity(intent)
-        }
-        // on pressing cancel button
-        alertDialog.setNegativeButton("Cancel") { dialog, which -> dialog.cancel() }
-        alertDialog.show()
-    }
-
-    override fun onLocationChanged(location: Location) {}
-    override fun onProviderDisabled(provider: String) {}
-    override fun onProviderEnabled(provider: String) {}
-    override fun onStatusChanged(provider: String, status: Int, extras: Bundle) {}
-    override fun onBind(arg0: Intent): IBinder? {
-        return null
-    }
-
-    companion object {
-
-        // The minimum distance to change Updates in meters
-        private val MIN_DISTANCE_CHANGE_FOR_UPDATES: Long = 10 // 10 meters
-        // The minimum time between updates in milliseconds
-        private val MIN_TIME_BW_UPDATES = (1000 * 60 * 1).toLong() // 1 minute
-    }
-}
