--- app/src/main/java/com/weatherkremenchug/weathermain/constants/Const.kt
@@ -1,5 +1,0 @@
-package com.weatherkremenchug.weathermain.constants
-
-// MainActivity
-const val APP_ID = "c3eebf803f44713f50e31a7c5b215a73"
-const val APP_PREFERENCES = "settings"
