--- app/src/main/java/com/weatherkremenchug/weathermain/data/CloudsData.kt
@@ -1,6 +1,0 @@
-package com.weatherkremenchug.weathermain.data
-
-class CloudsData {
-
-    val all: String? = null
-}
