--- app/src/main/java/com/weatherkremenchug/weathermain/data/CoordData.kt
@@ -1,8 +1,0 @@
-package com.weatherkremenchug.weathermain.data
-
-
-class CoordData {
-
-    val lat: String? = null
-    val lon: String? = null
-}
