--- app/src/main/java/com/weatherkremenchug/weathermain/data/ListData.kt
@@ -1,15 +1,0 @@
-package com.weatherkremenchug.weathermain.data
-
-import java.util.ArrayList
-
-class ListData {
-
-    val dt: String? = null
-    val main: MainWeatherData? = null
-    val weather = ArrayList<WeatherData>()
-    val clouds: CloudsData? = null
-    val wind: WindData? = null
-    val rain: RainData? = null
-    val sys: SysData? = null
-    val dt_txt: String? = null
-}
