--- app/src/main/java/com/weatherkremenchug/weathermain/data/MainWeatherData.kt
@@ -1,13 +1,0 @@
-package com.weatherkremenchug.weathermain.data
-
-class MainWeatherData {
-
-    val temp: String? = null
-    val temp_min: String? = null
-    val temp_max: String? = null
-    val pressure: String? = null
-    val sea_level: String? = null
-    val grnd_level: String? = null
-    val humidity: String? = null
-    val temp_kf: String? = null
-}
