--- app/src/main/java/com/weatherkremenchug/weathermain/data/RainData.kt
@@ -1,3 +1,0 @@
-package com.weatherkremenchug.weathermain.data
-
-class RainData
