--- app/src/main/java/com/weatherkremenchug/weathermain/data/SysData.kt
@@ -1,8 +1,0 @@
-package com.weatherkremenchug.weathermain.data
-
-class SysData {
-//    val pod: String? = null
-//    val country: String? = null
-//    val sunrise: String? = null
-//    val sunset: String? = null
-}
