--- app/src/main/java/com/weatherkremenchug/weathermain/data/WeatherData.kt
@@ -1,9 +1,0 @@
-package com.weatherkremenchug.weathermain.data
-
-class WeatherData {
-
-    val id: String? = null
-    val main: String? = null
-    val description: String? = null
-    val icon: String? = null
-}
