--- app/src/main/java/com/weatherkremenchug/weathermain/data/WeatherForecast.kt
@@ -1,10 +1,0 @@
-package com.weatherkremenchug.weathermain.data
-
-import java.util.ArrayList
-
-class WeatherForecast {
-    val cod: String? = null
-    val message: String? = null
-    val cnt: String? = null
-    val list = ArrayList<ListData>()
-}
