--- app/src/main/java/com/weatherkremenchug/weathermain/data/WeatherOnToday.kt
@@ -1,19 +1,0 @@
-package com.weatherkremenchug.weathermain.data
-
-import java.util.ArrayList
-
-class WeatherOnToday {
-
-    val coord: CoordData? = null
-    val weather = ArrayList<WeatherData>()
-    val base: String? = null
-    val main: MainWeatherData? = null
-    val clouds: CloudsData? = null
-    val dt: String? = null
-    val wind: WindData? = null
-    //val sys: SysData? = null
-
-    val id: String? = null
-    val name: String? = null
-    val cod: String? = null
-}
