--- app/src/main/java/com/weatherkremenchug/weathermain/data/WindData.kt
@@ -1,7 +1,0 @@
-package com.weatherkremenchug.weathermain.data
-
-class WindData {
-
-    val speed: String? = null
-    val deg: String? = null
-}
