--- app/src/main/java/com/weatherkremenchug/weathermain/enums/Enum.kt
@@ -1,2 +1,0 @@
-package com.weatherkremenchug.weathermain.enums
-
