--- app/src/main/java/com/weatherkremenchug/weathermain/extensions/ViewExt.kt
@@ -1,8 +1,0 @@
-package com.weatherkremenchug.weathermain.extensions
-
-import android.content.Context
-import android.widget.Toast
-
-fun Context.showToast(text: CharSequence) {
-    Toast.makeText(this, text, Toast.LENGTH_LONG).show()
-}
