--- app/src/main/java/com/weatherkremenchug/weathermain/network/Service.kt
@@ -1,2 +1,0 @@
-package com.weatherkremenchug.weathermain.network
-
