+++ app/src/main/res/layout/activity_settings.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:theme="@style/AppTheme"
-    tools:context=".activity.settingsactivity.SettingsActivity">
+    tools:context=".activity.settings.SettingsActivity">
 
     <include
         layout="@layout/app_bar_settings"
