+++ app/src/main/res/layout/activity_weather_on_today.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity.weatherontoday.WeatherOnTodayActivity">
+    tools:context=".activity.weathertoday.WeatherOnTodayActivity">
 
     <include
         layout="@layout/app_bar_weather_today"
