+++ app/src/main/res/layout/content_main.xml
@@ -7,7 +7,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context=".activity.mainactivity.MainActivity"
+    tools:context=".activity.main.MainActivity"
     android:background="@color/colorGray"
     tools:showIn="@layout/app_bar_main">
