--- app/src/test/java/com/weatherkremenchug/weathermain/ExampleUnitTest.kt
@@ -1,17 +1,0 @@
-package com.weatherkremenchug.weathermain
-
-import org.junit.Test
-
-import org.junit.Assert.*
-
-/**
- * Example local unit test, which will execute on the development machine (host).
- *
- * @see [Testing documentation](http://d.android.com/tools/testing)
- */
-class ExampleUnitTest {
-    @Test
-    fun addition_isCorrect() {
-        assertEquals(4, (2 + 2).toLong())
-    }
-}
